+++ src/components/Notes/index.ts
@@ -1,0 +1,4 @@
+export { StructuredNotesView } from "./StructuredNotes";
+export type { StructuredNotesProps } from "./StructuredNotes";
+export { NotesExport } from "./NotesExport";
+export type { NotesExportProps } from "./NotesExport";